Set database query end date
If the RDD isn’t cached, this database query may be executed twice,
once to build an index and once to read the data.  If new data has been
added between the two queries, you can end up with ids in the main
dataset but not in the index.  This causes
“org.apache.mahout.math.IndexException: Index -1 is outside allowable
range”